+++ src/patcher.rs
@@ -1,5 +1,5 @@
 use std::fs::{File, OpenOptions};
-use std::io::{self, Read, Write, Seek, SeekFrom};
+use std::io::{self, Read, Seek, SeekFrom, Write};
 
 pub struct Patcher;
 
@@ -8,7 +8,22 @@ impl Patcher {
         Patcher
     }
 
-    pub fn patch_bytes(&mut self, file_path: &str, pattern: &[u8], replacement: &[u8]) -> io::Result<()> {
+    /// Patching bytes in a file
+    ///
+    /// # Arguments
+    /// * `file_path` - The path to the file to be patched.
+    /// * `pattern` - The byte sequence to search for in the file.
+    /// * `replacement` - The byte sequence to replace the found pattern with.
+    ///
+    /// # Returns
+    ///
+    /// * `io::Result<()>` - Returns `Ok(())` if the patching is successful, otherwise returns an `io::Error`.
+    pub fn patch_bytes(
+        &mut self,
+        file_path: &str,
+        pattern: &[u8],
+        replacement: &[u8],
+    ) -> io::Result<()> {
         let mut file = OpenOptions::new().read(true).write(true).open(file_path)?;
         let mut buffer = vec![0; 4096];
         let mut position = 0;
@@ -37,28 +52,65 @@ impl Patcher {
         Ok(())
     }
 
+    /// Check if a pattern is present in a file
+    ///
+    /// # Arguments
+    /// * `file_path` - The path to the file to be checked.
+    /// * `pattern` - The byte sequence to search for in the file.
+    ///
+    /// # Returns
+    ///
+    /// * `io::Result<bool>` - Returns `Ok(true)` if the pattern is found, otherwise returns `Ok(false)`.
     pub fn is_pattern_present(&self, file_path: &str, pattern: &[u8]) -> io::Result<bool> {
         let mut file = File::open(file_path)?;
         let mut buffer = Vec::new();
         file.read_to_end(&mut buffer)?;
 
-        Ok(buffer.windows(pattern.len()).any(|window| window == pattern))
+        Ok(buffer
+            .windows(pattern.len())
+            .any(|window| window == pattern))
     }
 
+    /// Find the offset of a string in a file
+    ///
+    /// # Arguments
+    /// * `file_path` - The path to the file to be checked.
+    /// * `search_string` - The string to search for in the file.
+    ///
+    /// # Returns
+    /// * `io::Result<i64>` - Returns the offset of the string in the file if found, otherwise returns `-1`.
     pub fn find_string_offset(&self, file_path: &str, search_string: &str) -> io::Result<i64> {
         let mut file = File::open(file_path)?;
         let mut buffer = Vec::new();
         file.read_to_end(&mut buffer)?;
 
         let search_bytes = search_string.as_bytes();
-        if let Some(pos) = buffer.windows(search_bytes.len()).position(|window| window == search_bytes) {
+        if let Some(pos) = buffer
+            .windows(search_bytes.len())
+            .position(|window| window == search_bytes)
+        {
             Ok(pos as i64)
         } else {
             Ok(-1)
         }
     }
 
-    pub fn patch_bytes_by_offset(&mut self, file_path: &str, offset: i64, replacement_bytes: &[u8]) -> io::Result<()> {
+    /// Patching bytes in a file by offset
+    ///
+    /// # Arguments
+    /// * `file_path` - The path to the file to be patched.
+    /// * `offset` - The offset in the file where the patching should start.
+    /// * `replacement_bytes` - The byte sequence to replace the bytes at the specified offset with.
+    ///
+    /// # Returns
+    ///
+    /// * `io::Result<()>` - Returns `Ok(())` if the patching is successful, otherwise returns an `io::Error`.
+    pub fn patch_bytes_by_offset(
+        &mut self,
+        file_path: &str,
+        offset: i64,
+        replacement_bytes: &[u8],
+    ) -> io::Result<()> {
         let mut file = OpenOptions::new().write(true).open(file_path)?;
         file.seek(SeekFrom::Start(offset as u64))?;
         file.write_all(replacement_bytes)?;